+++ lib/src/languages/java/java11/java11_grammar.dart
@@ -208,7 +208,12 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
                 (v[0] as ASTModifiers?) ?? ASTModifiers.modifiersNone;
             var type = v[1];
             var name = v[2];
-            return ASTClassField(type, name, modifiers.isFinal);
+            return ASTClassField(
+              type,
+              name,
+              modifiers.isFinal,
+              modifiers: modifiers,
+            );
           });
 
   Parser<ASTClassField> classFieldDeclarationWithValue() =>
@@ -229,6 +234,7 @@ class Java11GrammarDefinition extends Java11GrammarLexer {
               name,
               expression,
               modifiers.isFinal,
+              modifiers: modifiers,
             );
           });